requirements.txt
PyQt6==6.6.1
PyQt6-WebEngine==6.6.0
requests==2.31.0
fastapi
uvicorn
websocket-client